requirements.txt
gitpython
jigsawstack
python-dotenv
